Move addResultCaseBody to BaseProcessorUnit
